+++ Obsolete/old_driver/sources/PriseAir.cpp
@@ -5,7 +5,7 @@
 
 
 //Constructeurs
-PriseAir::PriseAir(const std::string& modele, const char& rang, const unsigned short int& aerodynamisme):m_nom(modele), m_rang(rang), m_aerodynamisme(aerodynamisme)
+PriseAir::PriseAir(const std::string& modele, const char& rang, const int& aerodynamisme):m_nom(modele), m_rang(rang), m_aerodynamisme(aerodynamisme)
 {
 }
 
@@ -20,12 +20,12 @@ PriseAir::~PriseAir()
 
 
 //Methodes statiques
-PriseAir* PriseAir::chargerPriseAir(const unsigned short int& id)
+PriseAir* PriseAir::chargerPriseAir(const int& id)
 {
 	PriseAir* PriseAirCharge = 0; //PriseAir a creer
 	std::string var=""; //contient les lignes du fichier
 	std::string chemin = "Data/composants/priseair.cdx";
-	unsigned short int idActuel = id + 1; //indique l'id actuellement lu dans le fichier
+	int idActuel = id + 1; //indique l'id actuellement lu dans le fichier
 
 	std::ifstream engine(chemin.c_str());
 
@@ -50,7 +50,7 @@ PriseAir* PriseAir::chargerPriseAir(const unsigned short int& id)
 
 		if(idActuel==id)
 		{
-			unsigned short int curseur;
+			int curseur;
 			var.erase(0,var.find_first_of(";")+1);
 
 			curseur=var.find_first_of(";");
@@ -63,7 +63,7 @@ PriseAir* PriseAir::chargerPriseAir(const unsigned short int& id)
 
 			std::string sAerodynamisme=var;
 
-			unsigned short int aerodynamisme;
+			int aerodynamisme;
 
 			std::istringstream iss;
 			iss.str(sAerodynamisme);
@@ -86,14 +86,14 @@ PriseAir* PriseAir::chargerPriseAir(const unsigned short int& id)
 }
 
 
-void PriseAir::infoPriseAir(const unsigned short int& id, unsigned int& prix)
+void PriseAir::infoPriseAir(const int& id, unsigned int& prix)
 {
 	std::string var = ""; //contient les lignes du fichier
 	std::string chemin ="Data/composants/priseair.cdx";
 	std::ifstream engine(chemin.c_str());
 	std::istringstream iss;
-	unsigned short int idActuel = 0; //indique l'id actuellement lu dans le fichier
-	unsigned short int curseur;
+	int idActuel = 0; //indique l'id actuellement lu dans le fichier
+	int curseur;
 	char rang;
 
 	if(!engine)
@@ -133,14 +133,14 @@ void PriseAir::infoPriseAir(const unsigned short int& id, unsigned int& prix)
 }
 
 
-void PriseAir::infoPriseAir(const unsigned short int& id, std::string& modele, char& rang, unsigned short int& aero, unsigned int& prix)
+void PriseAir::infoPriseAir(const int& id, std::string& modele, char& rang, int& aero, unsigned int& prix)
 {
 	std::string var = ""; //contient les lignes du fichier
 	std::string chemin ="Data/composants/priseair.cdx";
 	std::ifstream engine(chemin.c_str());
 	std::istringstream iss;
-	unsigned short int idActuel = 0; //indique l'id actuellement lu dans le fichier
-	unsigned short int curseur;
+	int idActuel = 0; //indique l'id actuellement lu dans le fichier
+	int curseur;
 
 	if(!engine)
 	{
@@ -209,7 +209,7 @@ char PriseAir::getRang() const
 }
 
 
-unsigned short int PriseAir::getAerodynamisme() const
+int PriseAir::getAerodynamisme() const
 {
 	return m_aerodynamisme;
 }
