+++ Obsolete/old_driver/sources/PriseAir.h
@@ -18,20 +18,20 @@ class PriseAir
 		
 	public:
 	//Constructeurs
-	PriseAir(const std::string& modele, const char& rang, const unsigned short int& aerodynamisme);
+	PriseAir(const std::string& modele, const char& rang, const int& aerodynamisme);
 	
 	//Destructeur
 	~PriseAir();
 	
 	//Methodes statiques
-	static PriseAir* chargerPriseAir(const unsigned short int& id);
-	static void infoPriseAir(const unsigned short int& id, unsigned int& prix);
-	static void infoPriseAir(const unsigned short int& id, std::string& modele, char& rang, unsigned short int& aero, unsigned int& prix);	
+	static PriseAir* chargerPriseAir(const int& id);
+	static void infoPriseAir(const int& id, unsigned int& prix);
+	static void infoPriseAir(const int& id, std::string& modele, char& rang, int& aero, unsigned int& prix);	
 		
 	//Accesseurs et Mutateurs
 	std::string getNom() const;
 	char getRang() const;
-	unsigned short int getAerodynamisme() const;
+	int getAerodynamisme() const;
 
 	//Méthodes
 		//Aucune
@@ -43,7 +43,7 @@ class PriseAir
 	//Attributs
 	std::string m_nom;
 	char m_rang;
-	unsigned short int m_aerodynamisme;
+	int m_aerodynamisme;
 };
