+++ Obsolete/old_driver/sources/Spoiler.h
@@ -15,20 +15,20 @@ class Spoiler
 {
 	public:
 	//Constructeurs
-	Spoiler(const std::string& modele, const char& rang, const unsigned short int& aerodynamisme);
+	Spoiler(const std::string& modele, const char& rang, const int& aerodynamisme);
 	
 	//Destructeur
 	~Spoiler();
 	
 	//Méthodes statiques
-	static Spoiler* chargerSpoiler(const unsigned short int& id);
-	static void infoSpoiler(const unsigned short int& id, unsigned int& prix);
-	static void infoSpoiler(const unsigned short int& id, std::string& modele, char& rang, unsigned short int& aero, unsigned int& prix);
+	static Spoiler* chargerSpoiler(const int& id);
+	static void infoSpoiler(const int& id, unsigned int& prix);
+	static void infoSpoiler(const int& id, std::string& modele, char& rang, int& aero, unsigned int& prix);
 	
 	//Accesseurs et Mutateurs
 	std::string getNom() const;
 	char getRang() const;
-	unsigned short int getAerodynamisme() const;
+	int getAerodynamisme() const;
 	
 	//Méthodes
 		//Aucune
@@ -40,7 +40,7 @@ class Spoiler
 	//Attributs
 	std::string m_nom;
 	char m_rang;
-	unsigned short int m_aerodynamisme;
+	int m_aerodynamisme;
 };
